Revert "Started work on a mapper, currently does not map IdeaTypes correctly"
This reverts commit e98bae92ac7e5c96222a5763675c3288c0ec403e.

Revert mapper idea
