--- IdeaBox.Backend/IdeaBox.Backend.csproj
@@ -1,15 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\IdeaBox.Data\IdeaBox.Data.csproj" />
-    <ProjectReference Include="..\IdeaBox.Mapper\IdeaBox.Mapper.csproj" />
-    <ProjectReference Include="..\IdeaBox.Storage\IdeaBox.Storage.csproj" />
-  </ItemGroup>
-
-</Project>
