+++ IdeaBox.Data/IdeaBox.Data.csproj
@@ -7,7 +7,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\IdeaBox.Mapper\IdeaBox.Mapper.csproj" />
     <ProjectReference Include="..\IdeaBox.Storage\IdeaBox.Storage.csproj" />
   </ItemGroup>
