+++ IdeaBox.Data/Models/Types/IdeaTypeAttribute.cs
@@ -1,4 +1,9 @@
-﻿
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
 namespace IdeaBox.Data.Models.Types
 {
     [AttributeUsage(AttributeTargets.Class, AllowMultiple = false)]
