+++ IdeaBox.Data/Models/Types/Suggestion.cs
@@ -1,9 +1,6 @@
-﻿using IdeaBox.Mapper.Attributes;
-
-namespace IdeaBox.Data.Models.Types
+﻿namespace IdeaBox.Data.Models.Types
 {
     [IdeaType("suggestie")]
-    [MapClass(typeof(Archive.IdeaType))]
     public class Suggestion : BaseIdeaType
     { }
 }
