+++ IdeaBox.Logic/IdeaBox.Logic.csproj
@@ -8,8 +8,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\IdeaBox.Data\IdeaBox.Data.csproj" />
-    <ProjectReference Include="..\IdeaBox.Mapper\IdeaBox.Mapper.csproj" />
-    <ProjectReference Include="..\IdeaBox.Storage\IdeaBox.Storage.csproj" />
   </ItemGroup>
 
 </Project>
