+++ IdeaBox.Logic/JsonConverters/CustomDateTimeConverter.cs
@@ -2,7 +2,7 @@
 using System.Text.Json;
 using System.Text.Json.Serialization;
 
-namespace IdeaBox.Backend.JsonConverters
+namespace IdeaBox.Logic.JsonConverters
 {
     public class CustomDateTimeConverter : JsonConverter<DateTime>
     {
