+++ IdeaBox.Logic/Validators/RequiredIfSetAttribute.cs
@@ -1,18 +1,14 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace IdeaBox.Backend.Validators
+namespace IdeaBox.Logic.Validators
 {
     [AttributeUsage(AttributeTargets.Property, AllowMultiple = true)]
-    public class RequiredIfValueAttribute : ValidationAttribute
+    public class RequiredIfSetAttribute : ValidationAttribute
     {
         public string PropertyName { get; }
-        public object Value { get; }
 
-        public RequiredIfValueAttribute(string propertyName, object value)
-        {
-            PropertyName = propertyName;
-            Value = value;
-        }
+        public RequiredIfSetAttribute(string propertyName)
+            => PropertyName = propertyName;
 
         protected override ValidationResult? IsValid(object? value, ValidationContext validationContext)
         {
@@ -20,10 +16,10 @@ namespace IdeaBox.Backend.Validators
             var property = obj.GetType().GetProperty(PropertyName)
                 ?? throw new InvalidOperationException($"No property found in {obj.GetType().Name} named {PropertyName}.");
 
-            var otherVal = property.GetValue(obj);
-            if (Value.Equals(otherVal) && value == null)
-                return new ValidationResult($"The {validationContext.DisplayName} field must be set if {PropertyName} is set to \"{Value.ToString()}\".");
+            if (property.GetValue(obj) != null && value == null)
+                return new ValidationResult($"The {validationContext.DisplayName} field is required if {PropertyName} is set.");
             return ValidationResult.Success;
         }
+
     }
 }
