--- IdeaBox.Mapper/Attributes/MapClassAttribute.cs
@@ -1,12 +1,0 @@
-﻿
-namespace IdeaBox.Mapper.Attributes
-{
-    [AttributeUsage(AttributeTargets.Class, AllowMultiple = false)]
-    public class MapClassAttribute : Attribute
-    {
-        public Type Type { get; }
-
-        public MapClassAttribute(Type type)
-            => Type = type;
-    }
-}
