--- IdeaBox.Mapper/Attributes/MapPropertyAttribute.cs
@@ -1,15 +1,0 @@
-﻿
-using System.Reflection;
-
-namespace IdeaBox.Mapper.Attributes
-{
-    [AttributeUsage(AttributeTargets.Property, AllowMultiple = false)]
-    public class MapPropertyAttribute : Attribute
-    {
-        public string Name { get; }
-        public bool IsNested { get; init; }
-
-        public MapPropertyAttribute(string name)
-            => Name = name;
-    }
-}
