--- IdeaBox.Mapper/IMapper.cs
@@ -1,8 +1,0 @@
-﻿
-namespace IdeaBox.Mapper
-{
-    public interface IMapper
-    {
-        public object Map(object obj);
-    }
-}
