+++ IdeaBox.Storage/IStorageItem.cs
@@ -1,9 +1,10 @@
-﻿
-namespace IdeaBox.Storage
+﻿namespace IdeaBox.Storage
 {
     public interface IStorageItem
     {
         public int Id { get; set; }
         public DateTime CreationDate { get; set; }
+
+        public void Sanitise();
     }
 }
